Added a debugging switch to newrfi.csh that prevents the temp data from being deleted after completion of the program.
git-svn-id: https://svn.hcro.org/mmm@252 9b194dde-2b4d-0410-98f0-bba1668eb47e

+++ karto/RAPIDBeta/newrfi.csh
@@ -38,6 +38,7 @@ set nopol = "nopol" #Switch to apply polarization corrections to data
 set crosspol = "nocross" #Swtich to exclude XY and YX polarization - should be done if bandpass and pol correction information are not available
 set autoedge = 0
 set autoedgechan = 100
+set debug = 0
 
 varassign:
 
@@ -79,6 +80,8 @@ else if ("$argv[1]" =~ 'options='*) then
 	    set autoedge = 1
 	else if ($option == "noautoedge") then
 	    set autoedge = 0
+	else if ($option == "debug") then
+	    set debug = 1
 	else
 	    set badopt = ($badopt $option)
 	endif
@@ -360,13 +363,13 @@ end
 set times = (`date +%s.%N | awk '{print int(($1-date1)/60),int(($1-date1)%60)}' date1=$date1`)
 echo "Scanning process took $times[1] minute(s) $times[2] second(s)."
 
-rm -r $wd
+if !($debug) rm -r $wd
 
 finish:
 
 exit 0
 
 fail:
-rm -rf $wd
+if !($debug) rm -rf $wd
 echo "Houston, we have a problem..."
 exit 1
